+++ terraform/variables.tf
@@ -183,6 +183,10 @@ variable "inspec_wan_name" {
   default = "inspec-nw-wan"
 }
 
+variable "dataset_type" {
+  default = "CosmosDbSqlApiCollection"
+}
+
 variable "inspec_adls_file_system_name" {
   default = "inspec-adls-fs"
 }